Bring in an initial version of the FIT spec
Add the build infrastructure from devicetree.org since we want this spec
to be familiar to users of that spec.

Bring in some of the source from the U-Boot tree, with a small number of
updates to the initial paragraphs.

The definitions do not show as intended using this style. That can be
adjusted later.

Signed-off-by: Simon Glass <sjg@chromium.org>