Document how FIT should be used in a bootloader
The current documentation mentioned configuration matching but only in
passing and in a reference to U-Boot.

Add a new section which focusses on how FIT is used in the bootloader,
including more detailed information about configuration matching.

Adjust the material to ensure that it does not assume that the
bootloader is always U-Boot.

Signed-off-by: Simon Glass <sjg@chromium.org>