chapter2: add optional cmdline property to configuration
The Linux kernel needs a command line to indicate the root file system.
Further parameters may be passed on the command line [1].

Add an optional cmdline property to configuration nodes.

The UEFI specification uses UCS-2 encoded strings. UCS-2 is a deprecated
encoding of the Unicode Basic Multilingual Plane. Using UTF-8 for
encoding the cmdline property allows to represent all UCS-2 characters
when passing the string to an EFI binary via the EFI Loaded Image Protocol
as a load option.

[1] The kernel’s command-line parameters
    https://www.kernel.org/doc/html/v4.14/admin-guide/kernel-parameters.html

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>