Add support for logos
Provide a way to specify one or more logos so that the bootloader can
present a nicer graphical menu for the OS.

Support BMP as a baseline, with PNG and SVG as optional features.

Add a note into the usage section about boot menus, with some
recommendations for how a logo should be selected from the available
options.

Signed-off-by: Simon Glass <sjg@chromium.org>